reworking models to be a bit more flexible by allowing different error measures, and also separating the concerns so the model doesn't need to know about error

+++ core/core-math/src/main/java/org/openimaj/math/util/distance/ThresholdDistanceCheck.java
@@ -30,20 +30,22 @@
 package org.openimaj.math.util.distance;
 
 /**
- * Implementation of a {@link DistanceCheck} that tests the
- * distance against a fixed threshold.
+ * Implementation of a {@link DistanceCheck} that tests the distance against a
+ * fixed threshold.
  * 
  * @author Jonathon Hare (jsh2@ecs.soton.ac.uk)
- *
+ * 
  */
 public class ThresholdDistanceCheck implements DistanceCheck {
-	float threshold;
-	
+	double threshold;
+
 	/**
 	 * Construct with the given threshold for checking the distance.
-	 * @param threshold the threshold.
+	 * 
+	 * @param threshold
+	 *            the threshold.
 	 */
-	public ThresholdDistanceCheck(float threshold) {
+	public ThresholdDistanceCheck(double threshold) {
 		this.threshold = threshold;
 	}
